+++ src/search-commons/services/generic-sql.adapter.ts
@@ -622,7 +622,50 @@ export abstract class GenericSqlAdapter <R extends Record, F extends GenericSear
         return result;
     }
 
-    protected abstract extractTable(params: AdapterQuery): string;
+    protected extractTable(params: AdapterQuery): string {
+        if (params.where === undefined) {
+            return undefined;
+        }
+
+        let tabKey;
+        const types = params.where['type_txt'];
+        if (types !== undefined && types.in !== undefined) {
+            tabKey = this.extractSingleElement(types.in);
+            if (tabKey !== undefined) {
+                tabKey = tabKey.toLocaleLowerCase();
+                if (this.getTableConfigForTableKey(tabKey) !== undefined) {
+                    return tabKey;
+                }
+
+                return undefined;
+            }
+        }
+
+        const ids = params.where['id'];
+        if (ids !== undefined) {
+            tabKey = this.extractSingleElement(ids.in_number);
+            if (tabKey !== undefined) {
+                tabKey = tabKey.replace(/_.*/g, '').toLocaleLowerCase()
+                if (this.getTableConfigForTableKey(tabKey) !== undefined) {
+                    return tabKey;
+                }
+
+                return undefined;
+            }
+
+            tabKey = this.extractSingleElement(ids.in);
+            if (tabKey !== undefined) {
+                tabKey = tabKey.replace(/_.*/g, '').toLocaleLowerCase()
+                if (this.getTableConfigForTableKey(tabKey) !== undefined) {
+                    return tabKey;
+                }
+
+                return undefined;
+            }
+        }
+
+        return undefined;
+    }
 
     protected abstract getTableConfig(params: AdapterQuery): TableConfig;
 
@@ -677,5 +720,32 @@ export abstract class GenericSqlAdapter <R extends Record, F extends GenericSear
 
         return this.sqlQueryBuilder.queryTransformToAdapterWriteQuery(tableConfig, method, mappedProps, <AdapterOpts>opts);
     }
+
+    protected extractSingleElement(values: any): string {
+        if (values === undefined) {
+            return undefined;
+        }
+
+        if (!Array.isArray(values)) {
+            return values;
+        }
+
+        if (values.length === 1) {
+            return values[0];
+        }
+
+        const realValues = [];
+        values.map(value => {
+            if (value !== undefined && value.trim().length > 0) {
+                realValues.push(value.trim());
+            }
+        })
+        if (realValues.length === 1) {
+            return realValues[0];
+        }
+
+        return undefined;
+    }
+
 }